backend/requirements.txt
py-clob-client
fastapi
uvicorn
python-dotenv
requests
web3
